+++ src/app/knowledge-bases/modals/modal-detail-knowledge-base/modal-detail-knowledge-base.component.ts
@@ -10,9 +10,17 @@ export class ModalDetailKnowledgeBaseComponent implements OnInit {
   @Input() kb: KB;
   @Output() closeBaseModal = new EventEmitter();
   @Output() updateKnowledgeBase = new EventEmitter();
+
+  name: string;
+  source: string;
+  content: string;
+
   constructor() { }
 
   ngOnInit(): void {
+    this.name = this.kb.name;
+    this.source = this.kb.source;
+    this.content = this.kb.content;
   }
 
 
@@ -29,9 +37,11 @@ export class ModalDetailKnowledgeBaseComponent implements OnInit {
   }
 
   onUpdateKnowledgeBase(){
-    // console.log('onUpdateKnowledgeBase: ', this.kb);
+    this.kb.name = this.name;
+    this.kb.source = this.source;
+    this.kb.content = this.content;
+    //console.log('onUpdateKnowledgeBase: ', this.kb);
     this.updateKnowledgeBase.emit(this.kb);
   }
-
 
 }