+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # tiledesk-dashboard
 
+
+### 2.6.0
+- Deploys in production
+- Improves Knowledge Bases 
+- Adds the ability to create Knowledge Bases by adding a sitemap
+- Adds the customer.io integration
+- Fixes the bug: app cards are not all the same height
+
 ### 2.5.21
 - Deploys in production
 - Adds the ability to sort chatbot by name